Revised function and output
I made the main function calculate different regressions for all
possible combinations of the covariates and output a matrix of
coefficients and a vector of R2 values. But when I run the function, I
get the error that there is no inherited method. I don’t know how to
solve this problem. It must be coming from the getCoefs() and getR2s()
functions. In other words, I don’t know how to embed the functions I
defined in the main function.