Created subclass
I added a subclass called subRegressions, and the function
evalRegressions() works under the subclass. evalRegressions() uses
p-values to evaluate the importance of the coefficients. However, I
have not yet solved the issue I am having with making the functions
calculate different regressions for all possible combinations of the
covariates. I’ll come back to it later, because my functions do work
when I calculate a regression for all covariates.