Created show, plot, and print methods
I made the show, plot, and print methods for the main class
Regressions. But the problem is that the show and print methods seem to
have overridden the allRegressions() and evalRegressions() functions.
That is, when I run allRegressions() and evalRegressions(), I do not
get a matrix of coefficients and a vector of R2 values, but I get the
output of the print() function. This should be corrected easily, but I
don’t know how.